Updated the why support video component to handle videos for different country groups.
